+++ packages/formatters-util/src/util/registerWithMask.ts
@@ -1,4 +1,4 @@
-import type { ChangeEvent } from "react";
+import type { ChangeEvent, KeyboardEventHandler } from "react";
 import type { Path, PathValue, RegisterOptions, UseFormRegisterReturn, UseFormReturn } from "react-hook-form";
 import { formatFodselsnummer } from "../fodselsnummer/formatFodselsnummer";
 import { formatKontonummer } from "../kontonummer/formatKontonummer";
@@ -22,21 +22,67 @@ export type RegisterWithMaskOptions<T> = Omit<RegisterOptions<T>, "setValueAs">;
 const registerWithMask =
     (formatter: Formatter) =>
     <T>(form: UseFormReturn<T>, name: Path<T>, options?: RegisterWithMaskOptions<T>) => {
+        let onKeyDownCaretPosition = 0;
+        let onKeyDownKeyPressed = "";
+
         const setValueAs = (value: string) => value.replace(/\s/g, "");
         const onChange = (event: ChangeEvent<HTMLInputElement>) => {
             options?.onChange?.(event);
+
+            // save some values before event.target.value change
+            let onChangeCaretPosition = 0;
+            const inputLength = event.target.value.length;
+
+            // type checking formalities
+            if (event.target.selectionStart !== null) {
+                onChangeCaretPosition = event.target.selectionStart;
+            }
+
             form.setValue(
                 name as unknown as Path<T>,
                 formatters[formatter](event.target.value, { partial: true }) as PathValue<T, Path<T>>,
             );
+
+            let newPosition: number | null = null;
+
+            if (["Delete", "Backspace"].includes(onKeyDownKeyPressed)) {
+                // handle removing content
+                // calculate how much to move the caret, this also accounts for removing sections of text
+                const delta = onKeyDownCaretPosition - onChangeCaretPosition;
+
+                // calculate new caret position (- because we move backwards)
+                newPosition = onKeyDownCaretPosition - delta;
+            } else if (onChangeCaretPosition < event.target.value.length) {
+                // handle adding content from inside the string
+                // calculate how much to move the caret forwards
+                const delta = event.target.value.length - inputLength;
+
+                // calculate new caret position (+ because we move forwards)
+                newPosition = onChangeCaretPosition + delta;
+            }
+
+            if (newPosition !== null) {
+                event.target.setSelectionRange(newPosition, newPosition, undefined);
+            }
         };
+
         const register = form.register(name, { ...options, setValueAs, onChange });
-        let extra = {};
+
+        // save the caret position before the change occured
+        const onKeyDown: KeyboardEventHandler<HTMLInputElement> = (event) => {
+            if ((event.target as HTMLInputElement).selectionStart !== null) {
+                onKeyDownCaretPosition = (event.target as HTMLInputElement).selectionStart as number;
+            }
+            onKeyDownKeyPressed = event.key;
+        };
+
+        // add onKeyDown event handler to the registered input
+        const extra: Record<string, unknown> = {
+            onKeyDown,
+        };
 
         if (formatter === "number") {
-            extra = {
-                align: "right", // Se https://github.com/fremtind/jokul/pull/2898
-            };
+            extra.align = "right"; // Se https://github.com/fremtind/jokul/pull/2898
         }
 
         return Object.assign(register, extra);